64-bit Porting. JNI pointers passed in int.
BUG=354405, 346626

Review URL: https://codereview.chromium.org/263833002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267866 0039d316-1c4b-4281-b951-d872f2087c98

+++ components/web_contents_delegate_android/android/java/src/org/chromium/components/web_contents_delegate_android/ColorChooserAndroid.java
@@ -48,7 +48,7 @@ public class ColorChooserAndroid {
 
     @CalledByNative
     public static ColorChooserAndroid createColorChooserAndroid(
-            int nativeColorChooserAndroid,
+            long nativeColorChooserAndroid,
             ContentViewCore contentViewCore,
             int initialColor,
             ColorSuggestion[] suggestions) {
